Feat : 기업분석 API

See merge request s08-final/S08P31B107!61

+++ back/SourceCompany/SourceCompany/src/main/java/com/jobtang/sourcecompany/api/corp_detail/util/Calculator.java
@@ -15,21 +15,21 @@ public class Calculator {
     // a/b * 100
     public Double myRatio(Number a, Number b) {
         if (a == null || b == null) {return null;}
-        Double x = Double.valueOf(a.intValue());
-        Double y = Double.valueOf(b.intValue());
+        Double x = Double.valueOf(a.longValue());
+        Double y = Double.valueOf(b.longValue());
         return x / y * 100;
     }
 
     // a + b
     public Double myPlus(Number a, Number b) {
         if (a == null || b == null) {return null;}
-        Double x = Double.valueOf(a.intValue());
-        Double y = Double.valueOf(b.intValue());
+        Double x = Double.valueOf(a.longValue());
+        Double y = Double.valueOf(b.longValue());
         return x + y;
     }
 
     // a - b
-    public Integer mySubtraction(Integer a, Integer b) {
+    public Long mySubtraction(Long a, Long b) {
         if (a == null || b == null) {return null;}
         return a - b;
     }
@@ -37,45 +37,45 @@ public class Calculator {
     // a/(b+c)
     public Double myRatioWithSum(Number a, Number b, Number c) {
         if (a == null || b == null || c == null) {return null;}
-        Double x = Double.valueOf(a.intValue());
-        Double y = Double.valueOf(b.intValue());
-        Double z = Double.valueOf(c.intValue());
+        Double x = Double.valueOf(a.longValue());
+        Double y = Double.valueOf(b.longValue());
+        Double z = Double.valueOf(c.longValue());
         return x / (y + z) * 100;
     }
 
     // (a-b)/c
     public Double myRatioWithSubtraction(Number a, Number b, Number c) {
         if (a == null || b == null || c == null) {return null;}
-        Double x = Double.valueOf(a.intValue());
-        Double y = Double.valueOf(b.intValue());
-        Double z = Double.valueOf(c.intValue());
+        Double x = Double.valueOf(a.longValue());
+        Double y = Double.valueOf(b.longValue());
+        Double z = Double.valueOf(c.longValue());
         return (x - y) / z * 100;
     }
 
     // a * b
     public Double myMultiply(Number a, Number b) {
         if (a == null || b == null) {return null;}
-        Double x = Double.valueOf(a.intValue());
-        Double y = Double.valueOf(b.intValue());
+        Double x = Double.valueOf(a.longValue());
+        Double y = Double.valueOf(b.longValue());
         return x * y;
     }
 
     // a*(b+c)
     public Double myMultiplyWithSum(Number a, Number b, Number c) {
         if (a == null || b == null || c == null) {return null;}
-        Double x = Double.valueOf(a.intValue());
-        Double y = Double.valueOf(b.intValue());
-        Double z = Double.valueOf(c.intValue());
+        Double x = Double.valueOf(a.longValue());
+        Double y = Double.valueOf(b.longValue());
+        Double z = Double.valueOf(c.longValue());
         return x * (y + z);
 
     }
     // (a-b)/c * d
     public Double myCalculate(Number a, Number b, Number c, Number d) {
         if (a == null || b == null || c == null || d == null) { return null;}
-        Double x = Double.valueOf(a.intValue());
-        Double y = Double.valueOf(b.intValue());
-        Double z = Double.valueOf(c.intValue());
-        Double w = Double.valueOf(d.intValue());
+        Double x = Double.valueOf(a.longValue());
+        Double y = Double.valueOf(b.longValue());
+        Double z = Double.valueOf(c.longValue());
+        Double w = Double.valueOf(d.longValue());
         return (x - y) / z * w;
     }
 }
